Extract enterprise extension boundary
