+++ app/lib/glossia/extensions.ex
@@ -15,11 +15,31 @@ defmodule Glossia.Extensions do
     Application.get_env(:glossia, :authorizer, Glossia.Authorizers.Default)
   end
 
+  def blog do
+    Application.get_env(:glossia, :blog, Glossia.Blog.Default)
+  end
+
+  def docs do
+    Application.get_env(:glossia, :docs, Glossia.Docs.Default)
+  end
+
+  def features do
+    Application.get_env(:glossia, :features, Glossia.Features.Default)
+  end
+
+  def changelog do
+    Application.get_env(:glossia, :changelog, Glossia.Changelog.Default)
+  end
+
+  def marketing_router do
+    Application.get_env(:glossia, :marketing_router, Glossia.MarketingRouter.Default)
+  end
+
   def navigation do
     Application.get_env(:glossia, :navigation, Glossia.Navigation.Default)
   end
 
-  def account_nav_items(assigns) do
-    navigation().account_nav_items(assigns)
+  def account_nav_sections(sections, assigns) do
+    navigation().account_nav_sections(sections, assigns)
   end
 end
