+++ app/lib/glossia/marketing_router.ex
@@ -1,0 +1,5 @@
+defmodule Glossia.MarketingRouter do
+  @moduledoc false
+
+  @callback resolve(Plug.Conn.t(), [String.t()]) :: Plug.Conn.t() | :not_found
+end
