--- app/priv/blog/2026-02-03-why-l10n.md
@@ -1,62 +1,0 @@
-%{
-  title: "Localization was stuck in the past. We built Glossia to move it forward.",
-  summary: "Traditional localization tools add overhead, break CI, and lock you into vendor ecosystems. We're exploring what an agentic localization workflow can look like.",
-  date: ~D[2026-02-03],
-  slug: "2026-02-03-why-l10n",
-  author: "pedro"
-}
----
-
-If you've ever shipped software in more than one language, you know the drill. You pick a localization platform, connect it to your repository, and then spend the rest of your time managing the sync. Content goes out, translations come back, and somewhere in between things break.
-
-That overhead, the constant round-trip of content from and to your repository, is the tax every team pays for using today's localization tools. It sounds minor until you're the one debugging why a translation PR broke your site build at 6 PM on a Friday.
-
-## A design inherited from before the internet
-
-Most localization platforms were designed around concepts that predate the modern development workflow. Translation memories. Fuzzy matching. Human translators working inside proprietary editors, supported by tools that suggest similar strings from a database.
-
-These ideas made sense when translation was a manual, offline process. But companies turned translation memories into a lock-in mechanism. Your past translations, the institutional knowledge you paid for, live inside their platform. Moving to another provider means starting from scratch, or paying for an export that never quite works.
-
-The result is an industry built on artificial friction. Your content leaves your repo, enters a black box, and returns on someone else's schedule.
-
-## The broken feedback loop
-
-The problem is structural: external localization tools can't run your CI pipeline. They don't know about your linters, your build step, your link checker, or your frontmatter schema. They push translated content back to your repo and hope for the best. When it breaks, and it does, someone on the team has to stop what they're doing to fix formatting issues, broken syntax, or invalid markup that the translation tool introduced.
-
-LLMs and agentic experiences are presenting us with new opportunities to rethink these workflows entirely. An agent that generates a translation, runs your checks, sees the error, and retries until the output is valid. That kind of tight feedback loop changes everything.
-
-But it only works if the content stays where it lives: in your repository. The moment you send it off to an external platform, translations come back on someone else's timeline, and the integration breaks. The feedback that could have been instant now takes hours or days. The context that made it useful is long gone. You lose the loop, and with it, the entire advantage that agentic workflows were supposed to give you.
-
-## Observations that shaped Glossia
-
-These frustrations didn't turn into Glossia on their own. The project grew out of deep experience in both development and localization, which brought clarity to problems that are hard to see from just one side. Understanding the linguistic workflows, the human dynamics of translation teams, and the reasons why existing tools ended up the way they did was essential.
-
-Together, we kept arriving at the same observations: localization tooling was designed for a world without LLMs, without coding agents, and without CI pipelines. The entire model assumed that translation was something that happened outside the development workflow and got pushed back in. That made sense ten years ago. It doesn't anymore.
-
-We started asking: **what if localization agents could work the same way coding agents do?**
-
-We've been paying close attention to how [Anthropic](https://anthropic.com) thinks about agentic workflows with Claude. The pattern of giving an agent access to tools, letting it reason through a task, validate its own output, and iterate when something is off maps remarkably well to localization. A translation agent that can read your source files, understand project context, generate translations, run your linter, and fix issues before opening a pull request. That's not a fantasy. That's the workflow we're building.
-
-## Glossia is our gift to the software industry
-
-We built Glossia because we want more software to be localized, not less.
-
-Complicated processes and expensive platforms make localization inaccessible to small teams, indie developers, and side projects. If your translation workflow requires a procurement process, a per-word pricing negotiation, and a project manager to coordinate handoffs, most teams will just ship in English and call it a day.
-
-Glossia uses models you already have access to. And it validates output with your own tools, not ours.
-
-We think localization should be as natural as running your test suite.
-
-## An agent first, interfaces second
-
-At its core, Glossia is an agent. We're starting with the terminal as its primary interface because that's where the hardest problems get solved first: reading your source files, generating translations, running your checks, and iterating until the output is valid. This is the same pattern that [OpenAI](https://openai.com) followed with [Codex](https://openai.com/index/openai-codex/) and [Anthropic](https://anthropic.com) with [Claude Code](https://docs.anthropic.com/en/docs/claude-code). You build the agent, give it a terminal, and let it work.
-
-But the terminal is just the first interface, not the only one. We know that not everyone who contributes to localization quality is a developer. We talk about this often internally. The people who care most about translation accuracy, tone, and cultural nuance are often linguists and content specialists who don't think in terms of branches, compilation, or JSON.
-
-That's why we want to build new interfaces on top of the same agent. Something where a linguist sees the content, the context, and the translation side by side. They bring the human judgment that no model can replace. They refine what needs refining. And the agent handles everything else: committing, validating, opening the pull request.
-
-We don't have all the answers yet, and that's intentional. We'd rather build this thoughtfully than rush into a UI that misses the point. But the direction is clear: Glossia should welcome everyone who cares about making software speak every language.
-
-## Stay tuned
-
-Glossia is still early, and we're building it in the open. If any of this resonates with how you think about localization, keep an eye on the project. We'll be sharing more as we go.
