--- app/priv/changelog/2026-02-15-initial-release.md
@@ -1,16 +1,0 @@
-%{
-  title: "We are building Glossia",
-  summary: "We started working on Glossia and these are the first pieces we built."
-}
----
-
-We started working on Glossia, a platform for teams who want to manage multi-lingual and mono-lingual content with AI at the core. Here are the first pieces we built:
-
-## What we built so far
-
-- **Voice and glossary management** to keep your content consistent across every language.
-- **GitHub and GitLab integrations** so content updates flow straight from your repositories.
-- **Organization and member management** with role-based access control.
-- **A public REST API and MCP server** for programmatic access to all platform features.
-
-Follow this changelog for updates as we keep building.
