--- app/priv/changelog/2026-02-15-og-images.md
@@ -1,9 +1,0 @@
-%{
-  title: "Dynamic Open Graph images",
-  summary: "Link previews on social media now show page-specific images instead of a generic logo."
-}
----
-
-Sharing a Glossia link on Twitter/X, LinkedIn, Mastodon, or Slack now renders a rich preview card with the page title, description, and category. Images are generated lazily on the first request and cached on S3 so subsequent hits are fast.
-
-Every marketing page (blog posts, features, docs, changelog) and every public account or project page gets its own unique image. When content changes the hash changes, so the cache invalidates automatically.
