--- app/priv/changelog/2026-02-15-super-admin.md
@@ -1,6 +1,0 @@
-%{title: "Super admin system", summary: "Internal tooling for managing users, accounts, and access."}
----
-
-We added a super admin system for internal operations. Super admins can manage
-users, accounts, and access from a dedicated `/admin` area with its own layout
-and a private MCP server for programmatic administration.
