--- app/priv/changelog/2026-02-15-tickets.md
@@ -1,11 +1,0 @@
-%{
-  title: "Support tickets",
-  summary: "Report issues and request features directly from your dashboard. Each ticket includes a conversation thread where our team can help you."
-}
----
-
-We added a **Tickets** section to every account dashboard. You can now submit tickets to report issues or request features without leaving the app.
-
-Each ticket supports a full conversation thread so our team can ask for details and keep you updated on progress. Tickets have clear status labels (open, in progress, resolved, implemented) so you always know where things stand.
-
-Super admins can manage all tickets from the admin panel, reply as staff, and update ticket status. The same operations are available through the admin MCP server for integration with LLM tools.
