--- app/priv/changelog/2026-02-15-voice-about.md
@@ -1,13 +1,0 @@
-%{
-  title: "Voice: About section with target countries and cultural notes",
-  summary: "The Voice page now includes an About section where you describe what you do and pick target countries. AI-generated cultural notes help you tailor content for each market."
-}
----
-
-The Voice page has a new **About** section at the top. It includes:
-
-- A **description** field where you briefly explain what you do and who you serve.
-- A **target countries** picker with searchable country flags. Add as many countries as you need.
-- **Cultural notes** generated by AI for each selected country. When you fill in a description (20+ characters) and add at least one country, Glossia generates tailored cultural advice for communicating in that market. You can edit the generated notes before saving.
-
-All new fields are versioned alongside the rest of your voice settings, so you can track changes over time in the version history.
