--- app/priv/changelog/2026-02-17-voice-and-glossary-change-requests.md
@@ -1,14 +1,0 @@
-%{
-  title: "Voice and glossary suggestions for contributors",
-  summary: "Contributors can now submit structured voice and glossary suggestions as discussions, and authorized maintainers can apply them directly."
-}
----
-
-Voice and glossary collaboration now supports a proposal workflow:
-
-- Users without direct write access can edit the forms and submit a **suggestion** instead of saving immediately.
-- Suggestions are stored as structured discussions (`voice_suggestion` or `glossary_suggestion` kind).
-- Voice and Glossary pages now show open related suggestions.
-- Maintainers with write access can open a suggestion discussion and **apply it directly** from the discussion page.
-
-This keeps direct edits restricted while still letting external contributors propose concrete updates with full context.
