--- app/priv/changelog/2026-02-20-account-page-github-connect.md
@@ -1,7 +1,0 @@
-%{
-  title: "Account page for GitHub connections",
-  summary: "Moved GitHub connection controls under a new Account page accessible from the avatar menu."
-}
----
-
-You can now manage GitHub connections from a dedicated Account page in the avatar menu. The sidebar GitHub settings link has been removed to keep account management in one place.
